添加菜单模块；加入代码生成器工程

+++ MyProject/MyProject/MyProject.Application/Common/Dto/SysLSHInput.cs
@@ -38,7 +38,7 @@ namespace MyProject.Common.Dto
         /// <summary> 
         /// 流水号
         /// </summary> 
-        public Int64 Code { set; get; }
+        public int Code { set; get; }
 
 
         /// <summary>
